Inclui test scheduler de agodio y migre pid a 64 bits

+++ RowDaBoat-x64barebones-d4e1c147f975/Kernel/utilities/include/scheduler.h
@@ -7,25 +7,25 @@ uint64_t scheduler(uint64_t rsp);
 //El nombre de la funcion se encuentra como primer elemento de argv, pero esto no le llega al nuevo proceso.
 //stdFd debe ser un array de 2 elementos, el primero el fd de stdIn, y el segundo el de stdOut.
 //Tambien se permite que stdFd sea NULL, se tomaran ambos valores como 0.
-uint16_t initializeProccess(int (*function)(int , char **), uint8_t fg, int argc, char ** argv, uint16_t * stdFd);
+uint64_t initializeProccess(int (*function)(int , char **), uint8_t fg, int argc, char ** argv, uint16_t * stdFd);
 
-void kill(uint16_t pid);
+void kill(uint64_t pid);
 
-void block(uint16_t pid);
+void block(uint64_t pid);
 
-void unblock(uint16_t pid);
+void unblock(uint64_t pid);
 
 void exit();
 
-uint16_t getPID();
+uint64_t getPID();
 
 void dumpScheduler();
 
-void dumpProcessFromPID(uint16_t pid);
+void dumpProcessFromPID(uint64_t pid);
 
 void initScheduler();
 
-void changeProccessPriority(uint16_t pid, uint8_t prority);
+void changeProccessPriority(uint64_t pid, uint8_t prority);
 
 uint16_t getRunningProcessStdIn();
